backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Scraping & NLP
beautifulsoup4==4.12.2
newspaper3k==0.2.8
requests==2.31.0
lxml==4.9.3

# AI/ML
openai==1.3.7
anthropic==0.7.0
spacy==3.7.2

# Task Scheduling
apscheduler==3.10.4

# Email
sendgrid==6.11.0
# ou
resend==0.7.0

# Utils
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Dev tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1

feedparser==6.0.10
textblob==0.17.1

spacy==3.7.0
requests
feedparser
beautifulsoup4
rich
transformers
torch